Retry if an upsert fails due to a bad connection.
Sometimes the connection to the database server becomes invalid,
and this puts podyn into a mode where it continously emits errors
and fails to sync because it hangs on to the bad connection.

This change will create a new connection when this happens and
retry the command.

Only implemented on upsert for now, since that is by far the most
common operation.